Refine the last commit
Change table_half_steps, in fact it was exactly the same of full_steps, now ir is working in all encoder detents. I ve used the original table and now is working good, although it is not so bad to register steps just every two detents

Changed enc_handler_task stack size to 3072, sometimes very rapid movements resulted a stack overflow.

Deleted all references for flip_rotation, if it can be changed only in initialization, I can change the pins directly in the code too.

 On branch master
 Changes to be committed:
	modified:   components/encoder/encoder.c
	modified:   components/encoder/include/encoder.h
	modified:   main/main.c

+++ components/encoder/include/encoder.h
@@ -10,7 +10,7 @@ typedef struct encoder_s* encoder_handle_t;
 
 // Event structure for encoder output
 typedef struct {
-    int16_t steps; // Number of steps rotated (positive for CW, negative for CCW)
+    int32_t steps; // Number of steps rotated (positive for CW, negative for CCW)
     // Future additions: direction, accumulated value, etc.
 } encoder_event_t;
